+++ students/Victor_Iustus/Project/package.json
@@ -7,7 +7,7 @@
     "build": "webpack --mode development"
   },
   "keywords": [],
-  "author": "",
+  "author": "Victor Iustus",
   "license": "ISC",
   "devDependencies": {
     "css-loader": "^3.4.2",
