+++ students/Victor_Iustus/Project/src/public/index.html
@@ -4,8 +4,6 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="style/style.css">
-    <link rel="stylesheet" href="style/normalize.css">
     <title>Document</title>
 </head>
 <body>
@@ -19,17 +17,20 @@
                 </button>
             </form>
             <button class="btn-cart">Cart</button>
-            <div class="cart-block ">
+            <div id="cart-block" class="d-hidden">
                 <div class="d-flex">
                     <strong class="d-block">Всего товаров</strong> <div id="quantity"></div>
                 </div>
                 <hr>
                 <div class="cart-items">
-                    
+
                 </div>
                 <hr>
                 <div class="d-flex">
                     <strong class="d-block">Общая ст-ть:</strong> <div id="price"></div>
+                </div>
+                <div class="d-flex">
+                   <button name="cart-cleaner" id="cart-cleaner" class="btn -red margin-top-15">Очистить корзину</button>
                 </div>
             </div>
         </div>
@@ -37,8 +38,6 @@
     <main>
         <div class="products"></div>
     </main>
-    
     <script defer src="https://use.fontawesome.com/releases/v5.7.2/js/all.js" integrity="sha384-0pzryjIRos8mFBWMzSSZApWtPl/5++eIfzYmTgBBmXYdhvxPc+XcFEk+zJwDgWbP" crossorigin="anonymous"></script>
-    <script src="js/main.js"></script>
 </body>
 </html>
