+++ students/Victor_Iustus/Project/src/public/js/main.js
@@ -1,0 +1,9 @@
+import cart from './cart'
+import catalog from './catalog'
+
+
+export default function app() {
+ console.log('Jobs done!')
+ catalog.construct (cart) //тут происходит создание объекта и вся прочая магия
+ cart.constructor ()
+}
