+++ students/Victor_Iustus/Project/src/public/style/style.css
@@ -50,20 +50,28 @@ button:focus{
 p {
     margin: 0 0 5px 0;
 }
-.product-item{
+.product-item {
     display: flex;
     flex-direction: column;
+    justify-content: space-between;
     width: 200px;
+    min-height: 310px;
+    margin: 20px 0;
+    border: 1px solid #c0c0c040;
     border-radius: 5px;
     overflow: hidden;
-    margin: 20px 0;
+}
+.product-img {
+    width: 200px;
+    padding: 15px;
+    box-sizing: border-box;
 }
 img {
     max-width: 100%;
     height: auto
 }
 .desc {
-    border: 1px solid #c0c0c040;
+    border-top: 1px solid #c0c0c040;
     padding: 15px
 }
 
@@ -74,7 +82,7 @@ img {
 .cart{
     position: relative;
 }
-.cart-block{
+#cart-block{
     box-shadow: 0 0 5px rgba(0, 0, 0, 0.62);
     border-radius: 5px;
     box-sizing: border-box;
@@ -90,7 +98,7 @@ img {
 .invisible{
     display: none;
 }
-.cart-block:before{
+#cart-block:before{
     content: '';
     width: 0;
     height: 0;
@@ -102,21 +110,6 @@ img {
     border-bottom: 10px solid white;
 }
 
-.buy-btn, .del-btn{
-    margin-top: 5px;
-    background-color: #2f2a2d;
-    padding: 5px 15px;
-    border: 1px solid transparent;
-    color: #fafafa;
-    border-radius: 5px;
-    transition: all ease-in-out .4s;
-    cursor: pointer;
-}
-.buy-btn:hover, .del-btn:hover{
-    background-color: #fafafa;
-    color: #2f2a2d;
-    border: 1px solid #2f2a2d;
-}
 .cart-item {
     display: flex;
     justify-content: space-between;
@@ -139,7 +132,7 @@ img {
     margin-left: 30px;
 }
 .product-desc{
-    max-width: 150px;
+    flex: 1 0 auto;
 }
 .product-quantity {
     margin-top: 15px;
@@ -180,7 +173,40 @@ img {
     flex-direction: row;
     justify-content: space-between;
 }
-
+.d-hidden {
+    display: none
+}
 .d-block {
     display: block
+}
+.btn {
+    padding: 5px 15px;
+    border: 1px solid transparent;
+    border-radius: 5px;
+    transition: all ease-in-out .4s;
+    color: #000000;
+    cursor: pointer;
+
+}
+.btn.-black {
+    background-color: #2f2a2d;
+    color: #fafafa;
+}
+.margin-top-5 {
+    margin-top: 5px;
+}
+.margin-top-15 {
+    margin-top: 15px;
+}
+.btn.-black:hover {
+    background-color: #fafafa;
+    color: #2f2a2d;
+    border: 1px solid #2f2a2d;
+}
+.btn.-red {
+    background-color: red;
+    color: white;
+}
+.btn.-red:hover {
+    box-shadow: 0 5px 15px 0 rgba(227, 6, 17, .3);
 }
